+++ R/class-results.R
@@ -365,6 +365,7 @@ DiscreteTestResults <- R6Class(
     #' Prints a summary of the tested null hypotheses. The object itself is
     #' invisibly returned.
     #'
+    #' @importFrom checkmate assert_int qassert
     print = function(
       inputs = TRUE,
       pvalues = TRUE,
@@ -376,8 +377,10 @@ DiscreteTestResults <- R6Class(
       qassert(inputs, "B1")
       qassert(pvalues, "B1")
       qassert(supports, "B1")
-      qassert(test_idx, c("0", "X+[0,)"))
-      qassert(limit, "X1[0,)")
+      if(!is.null(test_idx) && !is.na(test_idx))
+        qassert(test_idx, "x+[0,)")
+      qassert(limit, c("0", "x1", "X1[0,)"))
+      assert_int(x = limit, na.ok = TRUE, lower = 0, null.ok = TRUE)
 
       if(inputs || pvalues){
         pars <- self$get_inputs(unique = FALSE)
@@ -389,13 +392,16 @@ DiscreteTestResults <- R6Class(
       cat("\n")
       cat(strwrap(private$test_name, prefix = "\t"), "\n")
       cat("\n")
-      cat("data:  ", private$data_name, "\n", sep = "")
+      cat("data: ", private$data_name, "\n")
+      cat("number of tests: ", length(private$p_values), "\n")
 
       if(any(inputs, pvalues, supports)) {
-        if(is.null(test_idx)){
-          n <- length(private$p_values)
+        n <- length(private$p_values)
+        if(is.null(test_idx)) {
+          limit <- ifelse(is.null(limit) || is.na(limit), n, limit)
           nums <- seq_len(ifelse(limit, min(limit, n), n))
-        } else nums <- test_idx
+        } else nums <- unique(pmin(na.omit(test_idx), n))
+
         for(i in nums) {
           cat("\n")
           cat("Test ", i, ":\n", sep = "")
@@ -466,15 +472,18 @@ DiscreteTestResults <- R6Class(
             print(supp[[i]], ...)
           }
         }
+        cat("\n")
+        if(is.null(test_idx) && limit > 0 && limit < n)
+          cat(
+            paste(
+              "[ print limit reached --", n - limit, "results omitted --",
+              "use print parameter 'limit' for more results ]"
+            )
+          )
+        if(!is.null(test_idx))
+          cat(paste("[", length(nums), "out of", n, "results printed ]"))
       }
       cat("\n")
-      if(is.null(test_idx) && limit > 0)
-        cat(
-          paste("[ print limit reached -- use print parameters 'test_idx'",
-                "or 'limit' for more results]"
-          ),
-          "\n"
-        )
 
       self
     }